pyproject.toml
[project]
name = "datarecovery"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "mutagen==1.47.0",
    "pathlib2>=2.3.7",
    "pillow==11.3.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["sort_music", "test_sort_music"]
